Fix graphql api address + use shared graphql client package + bump cloud-graphql version  (#14)
* Fix graphql api address + use shared graphql client package

* bump cloudgraphql
