+++ src/pkg/intents/genqlient.graphql
@@ -1,3 +1,3 @@
 mutation reportDiscoveredIntents($envId: ID!, $source: String!, $intents: [IntentInput!]!) {
-    reportDiscoveredSourcedIntents(environment: $envId, source: $source, intents: $intents)
+    reportDiscoveredSourcedIntents(environmentId: $envId, source: $source, intents: $intents)
 }
