+++ src/pkg/users/genqlient.yaml
@@ -1,8 +1,9 @@
 # genqlient config; for full documentation see:
 # https://github.com/Khan/genqlient/blob/main/docs/genqlient.yaml
 schema:
-  - ../../cloudgraphql/root.graphql
-  - ../../cloudgraphql/users.graphql
+  - ../../cloudgraphql/*.graphql
+  - ../../cloudgraphql/federation/*.graphql
+  - ../../cloudgraphql/deprecated/*.graphql
 operations:
 - genqlient.graphql
 generated: generated.go
